Add files via upload
I built a mini Student Management System using C language. This project is based on the concepts of arrays, and in this we can conveniently add, manage and modify the data record of students of a school or a university, like adding a new Student's record, modifying existing records, searching if a student exists or not and more. In this project, each task is distributed efficiently among different functions to improve the readability of the code and to help the reader better understand the concepts.